+++ web-ui/src/settings.ts
@@ -1,0 +1,5 @@
+export const modifyPlanLink = "https://github.com/marketplace/kodiakhq"
+export const installUrl = "https://github.com/marketplace/kodiakhq"
+export const docsUrl = "https://kodiakhq.com/docs/quickstart"
+export const helpUrl = "https://kodiakhq.com/help"
+export const activeUserUrl = "https://kodiakhq.com/docs/glossary#activeuser"
